Consolidate GenericAgentProxy
