+++ tests/test_onnx.py
@@ -3,8 +3,8 @@ import pytest
 import torch
 
 from gymnasium.vector import AsyncVectorEnv
+from emote.algorithms.generic_agent_proxy import GenericAgentProxy
 
-from emote.algorithms.sac import GenericAgentProxy
 from emote.extra.onnx_exporter import OnnxExporter
 from emote.nn.gaussian_policy import GaussianMlpPolicy as Policy
 
@@ -22,7 +22,11 @@ def exporter(tmpdir):
     policy = Policy(2, 1, [N_HIDDEN, N_HIDDEN])
 
     input_keys = list(env.dict_space.state.spaces.keys())
-    agent_proxy = GenericAgentProxy(policy, device, input_keys, ["actions"])
+    agent_proxy = GenericAgentProxy(policy, 
+                                    device, 
+                                    input_keys, 
+                                    True,
+                                    ["actions"])
 
     exporter = OnnxExporter(
         agent_proxy,
